hoan thanh chi tiet san pham

+++ app/src/main/java/com/example/bt_lon/activity/Detail_Activity.java
@@ -25,9 +25,11 @@ import android.widget.Toolbar;
 
 
 import com.example.bt_lon.R;
+import com.example.bt_lon.model.cart.Cart;
 import com.example.bt_lon.model.product.Product;
 import com.example.bt_lon.model.purchaseorder.PurchaseOrder;
 import com.example.bt_lon.model.user.RepositoryUser;
+import com.example.bt_lon.model.user.User;
 import com.example.bt_lon.sqlite_open_helper.DAO.CartDAO;
 import com.example.bt_lon.sqlite_open_helper.DAO.ProductDAO;
 import com.example.bt_lon.sqlite_open_helper.DAO.PurchaseOrderDAO;
@@ -146,31 +148,35 @@ public class Detail_Activity extends AppCompatActivity {
         btnMua.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                ProductDAO productDAO = new ProductDAO(Detail_Activity.this);
-                CartDAO cartDAO = new CartDAO(Detail_Activity.this);
-                PurchaseOrderDAO purchaseOrderDAO = new PurchaseOrderDAO(Detail_Activity.this);
-                int sl = Integer.parseInt(String.valueOf(btngiatri.getText()));
-                double cost = sl * product.getPrice();
-                PurchaseOrder purchaseOrder = new
-                        PurchaseOrder(
-                        product,
-                        RepositoryUser.getAccount(),
-                        sl,
-                        new Date(),
-                        cost
-                );
-                purchaseOrderDAO.insertPurchaseOrder(purchaseOrder);
-                if (sl == product.getQuantity()) {
-//                            productDAO.deleteProduct(cartList.get(i).getProduct().getProduct_id());
-                    productDAO.updateQuantity(product, 0);
-                    cartDAO.deleteProductFromCart(RepositoryUser.getAccount().getUser_id(), product.getProduct_id());
+                if (RepositoryUser.getAccount() != null) {
+                    ProductDAO productDAO = new ProductDAO(Detail_Activity.this);
+                    CartDAO cartDAO = new CartDAO(Detail_Activity.this);
+                    PurchaseOrderDAO purchaseOrderDAO = new PurchaseOrderDAO(Detail_Activity.this);
+                    int sl = Integer.parseInt(String.valueOf(btngiatri.getText()));
+                    double cost = sl * product.getPrice();
+                    PurchaseOrder purchaseOrder = new
+                            PurchaseOrder(
+                            product,
+                            RepositoryUser.getAccount(),
+                            sl,
+                            new Date(),
+                            cost
+                    );
+                    purchaseOrderDAO.insertPurchaseOrder(purchaseOrder);
+                    if (sl == product.getQuantity()) {
+                        productDAO.updateQuantity(product, 0);
+                        cartDAO.deleteProductFromCart(RepositoryUser.getAccount().getUser_id(), product.getProduct_id());
+                    } else {
+                        int quantityNew = product.getQuantity() - sl;
+                        productDAO.updateQuantity(product, quantityNew);
+                        cartDAO.deleteProductFromCart(RepositoryUser.getAccount().getUser_id(), product.getProduct_id());
+                    }
+                    finish();
                 } else {
-                    int quantityNew = product.getQuantity() - sl;
-                    productDAO.updateQuantity(product, quantityNew);
-                    cartDAO.deleteProductFromCart(RepositoryUser.getAccount().getUser_id(), product.getProduct_id());
+                    Intent intent = new Intent(Detail_Activity.this, LoginActivity.class);
+                    startActivity(intent);
                 }
-                Intent intent = new Intent(Detail_Activity.this, CartActivity.class);
-                startActivity(intent);
+
             }
         });
         dialog.show();
@@ -233,8 +239,21 @@ public class Detail_Activity extends AppCompatActivity {
         btnThem.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(Detail_Activity.this, CartActivity.class);
-                startActivity(intent);
+
+                if (RepositoryUser.getAccount() != null) {
+                    User user = RepositoryUser.getAccount();
+                    Cart cart = new Cart(user, product, 1);
+                    CartDAO cartDAO = new CartDAO(Detail_Activity.this);
+                    int sl = Integer.parseInt(String.valueOf(btngiatri.getText()));
+                    cartDAO.storeProductToCart(cart, sl);
+                    Intent intent = new Intent(Detail_Activity.this, CartActivity.class);
+                    startActivity(intent);
+                    dialog.dismiss();
+                    finish();
+                } else {
+                    Intent intent = new Intent(Detail_Activity.this, LoginActivity.class);
+                    startActivity(intent);
+                }
             }
         });
         dialog.show();
